Fix Iceberg deletion vector row marking and retries

+++ iceberg/common/src/main/scala/org/apache/iceberg/spark/source/GpuIcebergPartitionReader.scala
@@ -51,18 +51,29 @@ class GpuIcebergPartitionReader(private val task: GpuSparkInputPartition,
       file -> GpuDeleteFileInfo(scanTask.deletes().asScala.toSeq)
     }
   private lazy val deleteLoader = new DefaultDeleteLoader(rapidsFileIO, inputFiles, conf)
+  private lazy val projectsIsDeleted =
+    conf.expectedSchema.findField(MetadataColumns.IS_DELETED.fieldId()) != null
   private def deletionVectorProvider(file: IcebergPartitionedFile) = {
-    deleteInfoMap(file).deletionVector.map { delete =>
-      deleteLoader.loadDeletionVector(delete)
+    if (projectsIsDeleted) {
+      None
+    } else {
+      deleteInfoMap(file).deletionVector.map { delete =>
+        deleteLoader.loadDeletionVector(delete)
+      }
     }
   }
   private lazy val gpuDeleteFiterMap: Map[IcebergPartitionedFile, Option[GpuDeleteFilter]] =
     tasks.map {
       case (file, _) =>
         val postReadDeletes = deleteInfoMap(file).postReadDeletes
-        val filter = if (postReadDeletes.nonEmpty) {
+        val postReadDeletionVector = if (projectsIsDeleted) {
+          deleteInfoMap(file).deletionVector
+        } else {
+          None
+        }
+        val filter = if (postReadDeletes.nonEmpty || postReadDeletionVector.nonEmpty) {
           Some(new GpuDeleteFilter(rapidsFileIO, table.schema(),
-            inputFiles, conf, postReadDeletes))
+            inputFiles, conf, postReadDeletes, deletionVector = postReadDeletionVector))
         } else {
           None
         }
